+++ frontend/src/router/index.js
@@ -2,7 +2,7 @@ import Vue from "vue"
 import VueRouter from "vue-router"
 import Regist from "@/views/Regist"
 import RecruitmentRegister from "@/views/RecruitmentRegister"
-import LoginPage from "@/views/LoginPage"
+import Login from "@/views/Login"
 
 Vue.use(VueRouter)
 
@@ -17,8 +17,7 @@ const routes = [
   },
   {
     path: "/login",
-    name: "LoginPage",
-    component: LoginPage,
+    component: Login,
   },
 ]
